requirements.txt
flet==0.11.0
tensorflow==2.15.0
pillow==10.3.0
numpy==1.26.4
psycopg2-binary==2.9.9
